// frontend/src/Models/VehicleModel.ts
export class vehicleModel {
    "vehicle_id": number =0;
    "name_id": number = 0;
    "color": string= '';
    "vehicle_no": string= '';
    "rc": string= '';
    "km_driven": number =0;
    "last_service_km": number =0;
    "is_available": string= 'y';
}

export class vehicleNameModel 
{
    "name_id":number  = 0;
    "name": string  = '';
    "brand_id": number  = 0;
    "pph": number = 0;
    "kph": number = 0;
    "ekc": number = 0;
    "ehc": number = 0;
    "service_interval":number  = 0;
    "type_id": number = 0;
  }